chore: make functions available
